2. reflexion-agent/pyproject.toml
[project]
name = "2-reflexion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.30",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
